Fixed another migration file causing error

+++ db/migrate/20230509052843_drop_guests_table.rb
@@ -1,5 +1,9 @@
-class DropGuestsTable < ActiveRecord::Migration[7.0]
-  def change
-    drop_table :guests
+class DropGuestsTable < ActiveRecord::Migration[7.1]
+ def up
+    drop_table :guests if table_exists?(:guests)
+ end
+
+ def down
+    raise ActiveRecord::IrreversibleMigration
   end
 end
